shader: Fix caps truncation through .bam file

+++ panda/src/gobj/shaderModule.cxx
@@ -136,5 +136,5 @@ void ShaderModule::
 fillin(DatagramIterator &scan, BamReader *manager) {
   _stage = (Stage)scan.get_uint8();
   _source_filename = scan.get_string();
-  _used_caps = (int)scan.get_uint64();
+  _used_caps = scan.get_uint64();
 }
